Using Java 11 HttpClient instead of a reactive one.

+++ file-collector-application/file-collector-slave/src/main/java/com/github/filecollector/service/download/SourceDownloader.java
@@ -3,13 +3,13 @@ package com.github.filecollector.service.download;
 import com.github.filecollector.service.domain.DownloadTarget;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.core.io.buffer.DataBuffer;
-import org.springframework.core.io.buffer.DataBufferUtils;
 import org.springframework.stereotype.Service;
-import reactor.core.publisher.Flux;
-import reactor.core.publisher.Mono;
 
 import java.io.IOException;
+import java.net.http.HttpClient;
+import java.net.http.HttpRequest;
+import java.net.http.HttpResponse;
+import java.time.Duration;
 import java.util.Optional;
 
 @Slf4j
@@ -17,28 +17,36 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public class SourceDownloader {
 
-    private final DownloadRequestFactory downloadRequestFactory;
+    private final HttpClient client = HttpClient.newBuilder()
+            .connectTimeout(Duration.ofSeconds(10))
+            .followRedirects(HttpClient.Redirect.ALWAYS)
+            .build();
 
     public Optional<DownloadTarget> downloadToFile(final DownloadTarget downloadTarget) {
-        final Flux<DataBuffer> dataBufferFlux = downloadRequestFactory.newDownloadRequest(
-                downloadTarget.getSourceLocation());
+        log.info("Downloading: {}", downloadTarget.getSourceLocation().getLocation());
 
-        return DataBufferUtils.write(dataBufferFlux, downloadTarget.getTargetLocation().getPath())
-                .doOnError(error -> {
-                    try {
-                        downloadTarget.getTargetLocation().delete();
-                    } catch (final IOException e) {
-                        log.error("Failed to delete file on the staging location!", e);
-                    }
-                })
-                .then(Mono.just(downloadTarget))
-                .onErrorResume(error -> {
-                    if (log.isDebugEnabled()) {
-                        log.debug("Error downloading a document: {}!", error.getMessage());
-                    }
+        HttpRequest request = HttpRequest.newBuilder()
+                .uri(downloadTarget.getSourceLocation().getLocation())
+                .build();
 
-                    return Mono.empty();
-                })
-                .blockOptional();
+        try {
+            client.send(request, HttpResponse.BodyHandlers.ofFile(
+                    downloadTarget.getTargetLocation().getPath()));
+
+            return Optional.of(downloadTarget);
+        } catch (IOException | InterruptedException | IllegalArgumentException e) {
+            log.debug("Download failed for source: {} with reason: {}.",
+                    downloadTarget.getSourceLocation().getLocation(), e.getMessage());
+
+            try {
+                if (downloadTarget.getTargetLocation().exists()) {
+                    downloadTarget.getTargetLocation().delete();
+                }
+            } catch (final IOException ex) {
+                log.error("Failed to delete file on the staging location!", ex);
+            }
+
+            return Optional.empty();
+        }
     }
 }
